Move Asset and PriceRecord to latesco.core.data
Moved the Asset and PriceRecord class to latesco.core.data as it is a
more appropriate package. The Database interface and AssetTest files
have had their imports modified to reflect the change.
