+++ src/main/kotlin/latesco/core/data/PriceRecord.kt
@@ -15,14 +15,14 @@
  * along with Latesco.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package latesco.core
+package latesco.core.data
 
 import java.math.BigDecimal
 
 /**
  * Representation of an asset's price from a specific API at a given time.
  *
- * @param assetUid      latesco.core.Asset unique identifier which corresponds to this price record.
+ * @param assetUid      latesco.core.data.Asset unique identifier which corresponds to this price record.
  * @param apiUid        API unique identifier which corresponds to this price record.
  * @param timestamp     Timestamp when the price was fetched.
  * @param price         Price of the asset from a given API.
